refactor: request typename for discriminated union types in TS
