+++ shared/models/GetFirstAidLocations.ts
@@ -2,6 +2,7 @@ import { gql } from "@apollo/client";
 import { LocationSchema } from "./GetMapStrings";
 
 export type FirstAidSchema = {
+  __typename: "FirstAid";
   id: string;
   name: string;
   isWheelchairAccessible: boolean;
@@ -18,6 +19,7 @@ const query = (locale = "en") => {
   return gql`
       query {
         firstAids(locale:"${locale}", sort:"location.level_num:asc"){
+          __typename
           id
           name
           isWheelchairAccessible
