+++ shared/models/GetWaterFountainLocations.ts
@@ -3,6 +3,7 @@ import { gql } from "@apollo/client";
 import { LocationSchema } from "./GetMapStrings";
 
 export type WaterFountainSchema = {
+  typename: "WaterFountain";
   id: string;
   name: string;
   isWheelchairAccessible: boolean;
@@ -19,6 +20,7 @@ const query = (locale = "en") => {
   return gql`
       query {
         waterFountains(locale:"${locale}", sort:"location.level_num:asc"){
+          __typename
           id
           name
           isWheelchairAccessible
